Extract TableSelectHelper
