+++ java/clients/src/main/java/sleeper/clients/admin/PartitionsStatusReportScreen.java
@@ -20,31 +20,29 @@ import sleeper.configuration.properties.table.TableProperties;
 import sleeper.console.ConsoleInput;
 import sleeper.console.ConsoleOutput;
 import sleeper.console.UserExitedException;
-import sleeper.console.menu.ChooseOne;
 import sleeper.console.menu.Chosen;
 import sleeper.console.menu.ConsoleChoice;
 import sleeper.statestore.StateStoreException;
 import sleeper.status.report.partitions.PartitionsStatus;
 import sleeper.status.report.partitions.PartitionsStatusReporter;
 
-import static sleeper.clients.admin.AdminCommonPrompts.RETURN_TO_MAIN_MENU;
 import static sleeper.clients.admin.AdminCommonPrompts.confirmReturnToMainScreen;
 
 public class PartitionsStatusReportScreen {
     private final ConsoleOutput out;
     private final ConsoleInput in;
-    private final ChooseOne chooseOne;
+    private final TableSelectHelper tableSelectHelper;
     private final AdminConfigStore store;
 
     public PartitionsStatusReportScreen(ConsoleOutput out, ConsoleInput in, AdminConfigStore store) {
         this.out = out;
         this.in = in;
-        this.chooseOne = new ChooseOne(out, in);
+        this.tableSelectHelper = new TableSelectHelper(out, in);
         this.store = store;
     }
 
     public void chooseTableAndPrint(String instanceId) throws UserExitedException {
-        Chosen<ConsoleChoice> chosen = chooseTable();
+        Chosen<ConsoleChoice> chosen = tableSelectHelper.chooseTable();
         if (chosen.getChoice().isEmpty()) {
             TableProperties tableProperties = store.loadTableProperties(instanceId, chosen.getEntered());
             try {
@@ -56,19 +54,5 @@ public class PartitionsStatusReportScreen {
                 throw new RuntimeException(e);
             }
         }
-    }
-
-    private Chosen<ConsoleChoice> chooseTable() throws UserExitedException {
-        return chooseTable("")
-                .chooseUntilSomethingEntered(() ->
-                        chooseTable("\nYou did not enter anything please try again\n"));
-    }
-
-    private Chosen<ConsoleChoice> chooseTable(String message) {
-        out.clearScreen(message);
-        out.println("Which TABLE do you want to check?\n");
-        return chooseOne.chooseWithMessageFrom(
-                "Please enter the TABLE NAME now or use the following options:",
-                RETURN_TO_MAIN_MENU);
     }
 }
